wayland/gtk-shell: Add timestamp-less passive_present request
For when Gtk+ wants to present a surface but doesn't have a valid
timestamp (e.g. a non-interacted-with window got a new activate request
e.g. new file was opened), we don't to have Gtk+ have to fake a
timestamp to be able to ask to be presented. Thus, enable it to request
to be activated but without providing a timestamp. Currently this is
implemented by always calling "needs attention" on the corresponding
window.