kms/update: Make each MetaKmsUpdate update a single device
For now feedbacks from an update are combined, meaning we might loose
error information. The feedback API may have to be reconsidered and
redesigned when planes gets a more front seat position.

This means we need to avoid trying to post updates if we're in power
save mode, as it may be empty.